Add a "make test" rule for testing out the wheel file in a clean Linux container.
This is cobbled together from python-rpn continuous integration tests that are
being developed by Kristjan Onu.  See discussion at https://github.com/neishm/rmnlib-install-docker/issues/1

Note: Unlike the CI tests, the correctness of python-rpn is not being checked
here. The purpose of this test suite is to make sure the wheel is properly
packaged, with no missing modules / shared library symbols / etc.
This is currently only tested for a 64-bit Linux environment.

So far there are multiple failures (25 failures, 13 errors to be exact).  Still
need to determine which errors may be due to the wheel creation, and which ones
are upstream errors with python-rpn or the tests themselves.