Copy the libgfortran from the local archive, for all the manylinux wheels.
This is the same version used to compile the object files, so it should have
all the expected symbols.  The user shouldn't need to install libgfortran on
their system now.

We'll see how well this works - I think this is the fourth time I've tried to
embed the libgfortran dependencies???