Identify tests that are expected to fail in version 2.1.b3.
Also, use an older version of numpy for the tests, to avoid a whole bunch of
errors about comparing arrays to None.

Note: Individually, all the tests "pass" now (excluding the expected failures).
However, when all the tests are run together in a single pytest instance, there
are some additional failures from the vgd module that need to be sorted out.