Build wheels for Python-RPN version 2.1-u1.rc11.
This is re-versioned as "2.1rc11" for compatibility with pip.

Updates were merged in from the "pip-install" branch of my python-rpn fork.

Note: The latest test data has not been uploaded to a public repository.
If you have access to the "U1" machines, you could transfer the file
/home/min000/devel/rpnpy-tests/dist/eccc_rpnpy_tests-2.1rc11.zip
to your local wheelhouse/ directory.  Then you should be able to run "make test".