Updates for building and testing a macos binary wheel for fstd2nc-deps.
This is a squashed commit from all the attempts to get this working.
Apologies if you already checked out that line of commit history, you'll need to
rebase :(

For the full commit history that went into this, I might push it to a separate
branch for posterity (maybe "fstd2nc-macos-fun").